requirements.txt
pyyaml
PyQt5
pytest